+++ src/test/resources/db/component/component_link_by_id.sql
@@ -1,4 +1,4 @@
-PREPARE component_link_by_id (int, int) AS
+PREPARE component_link_by_id (bigint, bigint) AS
     SELECT     s.component_link_id,
                u.component_link_id AS user_component_link_id,
                s.user_id,
