+++ src/test/resources/db/component/component_link_by_max_pos.sql
@@ -1,4 +1,4 @@
-PREPARE component_link_by_max_pos (int, int) AS
+PREPARE component_link_by_max_pos (bigint, bigint) AS
     SELECT max(g.order_nbr) AS max_order_nbr
       FROM ( SELECT CASE WHEN (u.order_nbr IS NULL) THEN s.order_nbr ELSE u.order_nbr END AS order_nbr
                FROM component_links s
