+++ src/test/resources/db/log/change_log_named_by_field_row.sql
@@ -1,4 +1,4 @@
-PREPARE change_log_named_by_field_row (int, int, int, int) AS
+PREPARE change_log_named_by_field_row (bigint, bigint, bigint, bigint) AS
     SELECT s.change_id,
            s.user_id,
            s.change_time,
