+++ src/test/resources/db/result/result_by_id.sql
@@ -1,4 +1,4 @@
-PREPARE result_by_id (int) AS
+PREPARE result_by_id (bigint) AS
     SELECT group_id,
            formula_id,
            user_id,
