+++ src/test/resources/db/term/term_list_by_name_like.sql
@@ -1,4 +1,4 @@
-PREPARE term_list_by_name_like (int,text) AS
+PREPARE term_list_by_name_like (bigint,text) AS
     SELECT s.term_id,
            u.term_id AS user_term_id,
            s.user_id,
