+++ src/test/resources/db/triple/triple_list_by_phr_and_vrb_down.sql
@@ -1,4 +1,4 @@
-PREPARE triple_list_by_phr_and_vrb_down (int, int, int) AS
+PREPARE triple_list_by_phr_and_vrb_down (bigint, bigint, bigint) AS
     SELECT s.triple_id,
            u.triple_id AS user_triple_id,
            s.user_id,
