+++ src/test/resources/db/value/value_time_series_std_by_id.sql
@@ -1,4 +1,4 @@
-PREPARE value_time_series_std_by_id (int) AS
+PREPARE value_time_series_std_by_id (bigint) AS
     SELECT value_time_series_id,
            user_id,
            group_id,
